Give TypedAbiDeclaration and TypedTraitDeclaration their own files.
